Refactor Private.js file to Private.tsx file
